Add more error check into delagent unit test.
git-svn-id: http://svn.code.sf.net/p/fossology/code/branches/fossology2.0/fossology@4649 a5643f74-2041-0410-a2b9-94ffef304a07

+++ src/delagent/agent_tests/Unit/testCleanDB.sh
@@ -34,6 +34,11 @@ echo 'send "fossy\r"' >> $HOME/connectdb.exp
 echo 'interact' >> $HOME/connectdb.exp
 
 expect $HOME/connectdb.exp
+if [ $? -ne 0 ];
+then
+  echo "Delete Test database Error!"
+  exit 1
+fi
 rm -f $HOME/connectdb.exp
 
 if [ $? -ne 0 ];
@@ -44,7 +49,8 @@ fi
 
 PREFIX=`cat $VARS|grep -i '^PREFIX'|awk -F = '{print $2}'`
 cp $PREFIX/etc/fossology/Db.conf ~/
-#sed -i 's/fossologytest;/fossology;/' ~/Db.conf
+sed -i 's/fossologytest;/fossology;/' ~/Db.conf
 cp ~/Db.conf $PREFIX/etc/fossology/Db.conf
+rm -f ~/Db.conf
 
 echo "Clean Test database success!"
